+++ src/base_cf.cpp
@@ -36,7 +36,7 @@ void MGL_EXPORT mgl_set_flag(HMGL gr, int val, long flag)	{	gr->set(val,flag);	}
 //-----------------------------------------------------------------------------
 void MGL_EXPORT mgl_set_color(char id, double r, double g, double b)
 {
-	for(size_t i=0;mglColorIds[i].id;i++)
+	for(long i=0;mglColorIds[i].id;i++)
 		if(mglColorIds[i].id==id)	mglColorIds[i].col = mglColor(r,g,b);
 }
 void MGL_EXPORT mgl_set_color_(char *id, mreal *r, mreal *g, mreal *b, int)	{	mgl_set_color(*id,*r,*g,*b);	}
